backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.9.0
exceptiongroup==1.2.0
firebase-admin==6.4.0
Flask==3.0.2
Flask-Cors==4.0.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
langcodes==3.3.0
MarkupSafe==2.1.5
msgpack==1.0.7
openai==1.12.0
opencv-python-headless==4.10.0.84
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
tempfile2==0.1.2
tqdm==4.66.2
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.1
